AICTE/backend/requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# NumPy (use latest compatible version for Python 3.13)
numpy>=1.26.0

# ML/AI dependencies
transformers>=4.35.0
torch>=2.1.0
torchaudio>=2.1.0
openai-whisper

# Audio processing
ffmpeg-python>=0.2.0

